Tried fixing the csrf token issue
The backend and the frontend BOTH save the csrf-token as a cookie, but the frontend is having problems sending a POST request to the backend and getting a forbidden error (403) every time the front end makes a POST request. This is starting to get frustrating but I will keep trying to find a fix.