Refactored seedData.ts and updated registration
I refactored seedData.ts so it no longer clears all the data in the data base and inserts new data, now it updates the existing data in the database and adds new data if the data isn't present in the database.  I updated the registration html page so that Amanda's credentials check that she wrote functions properly in the registration page. Now the user cannot successfully register unless their email and password are valid. Onto login functionality next.