Reset Password Functionality Implemented
The user can now click the forgot password link on the login page and change their password once they enter their email and the new password. Once validation emails are implemented, I could change it so that it requires the validation code from that email, but that's a future me problem.