Attempt to implement search functionality
The routes for the server aren't being added to express correctly despite the setup being as intended, I honestly don't know how to fix the issue as of now. This is preventing the search functionality from working correctly because it can't find the /api/organizations/search route because it isn't being created and added to the routes that are used by express.