Attempt of Live Update functionality for Cards
I attempted to make it so that when a user is logged in, the cards in their corresponding sections are displayed properly. But I think it's not getting the userID from the right place so I need to figure out where that is stored or find a way to store it in the sessionStorage.