At my wits end
I tried to get the responsive cards to work, but every method I'm trying is just not working. I desperately need help to figure out how to solve this. Someone please contact a professor that is knowledgeable on the subject and is willing to help me out. I am going to stop working on this for now and try to work on the search system tomorrow.