Minor Changes to Nav Bar Pathing
The "My Orgs" section of the Nav Bar now leads to the the UserView.html page. The UserView.html page is planned to have a list of all the organizations the user has supported or followed, along with information about each organization and a way to reach the organization's website.