Fixed Nav Bar Destinations
The My organization tab now leads to the My Organization page. I'm planning on removing the "New Org" tab for normal users as they shouldn't be able to create any new organizations. There might be a need to create a page that allows admin users to create new organizations with a form, and then the data from the form is uploaded to the data base.