Small changes to the CreateSR page and the MemberView Page
The CreateSR page now has the filters labeled and spelled correctly, and I decided to get rid of the metrics section of the MemberView page as we won't be using that any more. The MemberView page now also displays the information of the organization that the member is affiliated with.