Changed admin nav bar
The admin nav bar now has the home, my organizations, profile, volunteer, and search tabs. The volunteer tab takes you to the normal homepage for normal users, and the search page takes you to the search page. My organizations now displays the organization that an Admin is apart of, along with and organizations associated with the events the user has signed up for or completed.